Encapsulate ffmpeg commands inside FfmpegService
Move image resize and audio trim logic directly into FfmpegService with
domain-specific public methods, making the run() helper and ffmpeg args
private. Delete the now-redundant AudioTrimService and ImageResizeService.
