--- server/src/main/java/io/github/mucsi96/learnlanguage/service/AudioTrimService.java
@@ -1,27 +1,0 @@
-package io.github.mucsi96.learnlanguage.service;
-
-import java.io.IOException;
-import java.util.List;
-
-import org.springframework.stereotype.Service;
-
-import lombok.RequiredArgsConstructor;
-
-@Service
-@RequiredArgsConstructor
-public class AudioTrimService {
-
-  private final FfmpegService ffmpegService;
-
-  public byte[] trimSilence(byte[] audioData) throws IOException {
-    return ffmpegService.process(List.of(
-        "ffmpeg", "-y", "-loglevel", "error",
-        "-f", "s16le", "-ar", "44100", "-ac", "1",
-        "-i", "pipe:0",
-        "-filter:a", "silenceremove=start_periods=1:start_threshold=-50dB:stop_periods=-1:stop_threshold=-50dB",
-        "-codec:a", "libmp3lame", "-b:a", "128k",
-        "-f", "mp3",
-        "pipe:1"
-    ), audioData);
-  }
-}
